+++ src/main/java/com/backend/immilog/user/application/dto/UserInfoUpdateCommand.java
@@ -1,4 +1,4 @@
-package com.backend.immilog.user.application.command;
+package com.backend.immilog.user.application.dto;
 
 import com.backend.immilog.user.model.enums.UserCountry;
 import com.backend.immilog.user.model.enums.UserStatus;
