+++ src/main/java/com/backend/immilog/user/application/services/EmailService.java
@@ -1,11 +1,11 @@
 package com.backend.immilog.user.application.services;
 
 import com.backend.immilog.user.exception.UserException;
-import com.backend.immilog.user.model.services.EmailService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 import javax.mail.MessagingException;
@@ -16,10 +16,10 @@ import static com.backend.immilog.post.exception.PostErrorCode.EMAIL_SEND_FAILED
 @Slf4j
 @RequiredArgsConstructor
 @Service
-public class EmailServiceImpl implements EmailService {
+public class EmailService {
     private final JavaMailSender javaMailSender;
 
-    @Override
+    @Async
     public void sendHtmlEmail(
             String to,
             String subject,
