+++ src/main/java/com/backend/immilog/user/model/entities/User.java
@@ -3,7 +3,7 @@ package com.backend.immilog.user.model.entities;
 import com.backend.immilog.user.model.enums.UserCountry;
 import com.backend.immilog.global.enums.UserRole;
 import com.backend.immilog.global.model.BaseDateEntity;
-import com.backend.immilog.user.application.command.UserSignUpCommand;
+import com.backend.immilog.user.application.dto.UserSignUpCommand;
 import com.backend.immilog.user.model.embeddables.Location;
 import com.backend.immilog.user.model.embeddables.ReportInfo;
 import com.backend.immilog.user.model.enums.UserStatus;
