--- src/main/java/com/backend/immilog/user/model/services/EmailService.java
@@ -1,12 +1,0 @@
-package com.backend.immilog.user.model.services;
-
-import org.springframework.scheduling.annotation.Async;
-
-public interface EmailService {
-    @Async
-    void sendHtmlEmail(
-            String to,
-            String subject,
-            String htmlBody
-    );
-}
