+++ src/main/java/com/backend/immilog/user/presentation/controller/AuthController.java
@@ -2,9 +2,9 @@ package com.backend.immilog.user.presentation.controller;
 
 import com.backend.immilog.global.presentation.response.ApiResponse;
 import com.backend.immilog.global.security.ExtractUserId;
+import com.backend.immilog.user.application.services.LocationService;
+import com.backend.immilog.user.application.services.UserSignInService;
 import com.backend.immilog.user.model.dtos.UserSignInDTO;
-import com.backend.immilog.user.model.services.LocationService;
-import com.backend.immilog.user.model.services.UserSignInService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
