+++ src/main/java/com/backend/immilog/user/presentation/request/UserInfoUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.backend.immilog.user.presentation.request;
 
-import com.backend.immilog.user.application.command.UserInfoUpdateCommand;
+import com.backend.immilog.user.application.dto.UserInfoUpdateCommand;
 import com.backend.immilog.user.model.enums.UserCountry;
 import com.backend.immilog.user.model.enums.UserStatus;
 import io.swagger.annotations.ApiModel;
