+++ src/main/java/com/backend/immilog/user/presentation/request/UserReportRequest.java
@@ -1,6 +1,6 @@
 package com.backend.immilog.user.presentation.request;
 
-import com.backend.immilog.user.application.command.UserReportCommand;
+import com.backend.immilog.user.application.dto.UserReportCommand;
 import com.backend.immilog.user.enums.ReportReason;
 import io.swagger.annotations.ApiModel;
 import lombok.Builder;
