+++ src/main/java/com/backend/immilog/user/presentation/request/UserSignUpRequest.java
@@ -1,6 +1,6 @@
 package com.backend.immilog.user.presentation.request;
 
-import com.backend.immilog.user.application.command.UserSignUpCommand;
+import com.backend.immilog.user.application.dto.UserSignUpCommand;
 import io.swagger.annotations.ApiModel;
 import lombok.Builder;
